Branchen gemerged, erster Versuch ListView
